(fix): Separated AuthProvider & PrivateRouting into separate folders
