+++ src/components/PrivateRoute/PrivateRoute.tsx
@@ -1,0 +1,7 @@
+import { Outlet, Navigate } from "react-router-dom";
+import { useAuth } from "../../hooks/useAuth";
+
+export const PrivateRoute = () => {
+  const { isAuthenticated } = useAuth();
+  return isAuthenticated ? <Outlet /> : <Navigate to="/error" replace />;
+};
